requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
pytest
flask-bcrypt
python-dotenv
mysqlclient
cryptography